refactor(concert): 관련 공연 조회 의존 방향 수정
