+++ src/main/java/com/backend/allreva/module/concert/concert/domain/ConcertRepository.java
@@ -1,6 +1,6 @@
 package com.backend.allreva.module.concert.concert.domain;
 
-import com.backend.allreva.module.concert.place.application.dto.RelatedConcertResponse;
+import com.backend.allreva.module.concert.concert.application.dto.RelatedConcertResponse;
 import java.util.List;
 import java.util.Optional;
