+++ src/main/java/com/backend/allreva/module/concert/concert/presentation/ConcertController.java
@@ -3,15 +3,20 @@ package com.backend.allreva.module.concert.concert.presentation;
 import com.backend.allreva.common.web.response.Response;
 import com.backend.allreva.module.concert.concert.application.ConcertService;
 import com.backend.allreva.module.concert.concert.application.dto.ConcertDetailResponse;
+import com.backend.allreva.module.concert.concert.application.dto.RelatedConcertResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Positive;
+import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RequiredArgsConstructor
@@ -28,5 +33,13 @@ public class ConcertController implements ConcertControllerSwagger {
     public Response<ConcertDetailResponse> findConcertDetail(@PathVariable("concertCode") final String concertCode) {
         ConcertDetailResponse detail = concertService.findDetailById(concertCode);
         return Response.onSuccess(detail);
+    }
+
+    @GetMapping("/by-hall")
+    public Response<List<RelatedConcertResponse>> findRelatedConcertsByHall(
+            @NotBlank @RequestParam final String hallCode,
+            @RequestParam(required = false) final String lastConcertCode,
+            @Positive @RequestParam(defaultValue = "3") final int pageSize) {
+        return Response.onSuccess(concertService.findRelatedConcertsByHall(hallCode, lastConcertCode, pageSize));
     }
 }
