+++ src/main/java/com/backend/allreva/module/concert/place/exception/ConcertHallErrorCode.java
@@ -9,8 +9,7 @@ import org.springframework.http.HttpStatus;
 @RequiredArgsConstructor
 public enum ConcertHallErrorCode implements ErrorCode {
     CONCERT_HALL_NOT_FOUND(HttpStatus.NOT_FOUND.value(), "CONCERT_HALL_NOT_FOUND", "존재하지 않는 공연장입니다"),
-    CONCERT_HALL_SEARCH_NOTFOUND(HttpStatus.NOT_FOUND.value(), "CONCERT_HALL_SEARCH_NOT_FOUND", "존재하는 공연장이 없습니다"),
-    RELATED_CONCERT_EXCEPTION(HttpStatus.INTERNAL_SERVER_ERROR.value(), "RELATED_CONCERT_EXCEPTION", "연관 콘서트 에러");
+    CONCERT_HALL_SEARCH_NOTFOUND(HttpStatus.NOT_FOUND.value(), "CONCERT_HALL_SEARCH_NOT_FOUND", "존재하는 공연장이 없습니다");
 
     private final int status;
     private final String code;
