+++ src/main/java/com/backend/allreva/module/concert/place/presentation/ConcertHallController.java
@@ -3,8 +3,6 @@ package com.backend.allreva.module.concert.place.presentation;
 import com.backend.allreva.common.web.response.Response;
 import com.backend.allreva.module.concert.place.application.HallService;
 import com.backend.allreva.module.concert.place.application.dto.ConcertHallDetailResponse;
-import com.backend.allreva.module.concert.place.application.dto.RelatedConcertResponse;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
@@ -20,11 +18,5 @@ public class ConcertHallController implements ConcertHallControllerSwagger {
     public Response<ConcertHallDetailResponse> findHallDetailByHallCode(final String hallCode) {
         ConcertHallDetailResponse details = hallService.findDetailByHallCode(hallCode);
         return Response.onSuccess(details);
-    }
-
-    @Override
-    public Response<List<RelatedConcertResponse>> findRelatedConcertList(
-            final String hallCode, final String lastConcertCode, final int pageSize) {
-        return Response.onSuccess(hallService.getRelatedConcert(hallCode, lastConcertCode, pageSize));
     }
 }
